Add files via upload
I a tire factory many workers are working there each worker is assigned with a integer number which help them to find their Supervisor. the project Manager having an Array that consist of a integer numbers help the project Manager to find their supervisors. Supervisor will be assigned as the first largest integer find on the right side of the array if not found return -1